requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-heroku==0.3.1
djangorestframework==3.14.0
h5py==3.9.0
numpy==1.25.2
pandas==2.0.3
psycopg2==2.9.7
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
